requirements.txt
catboost
cpbd==1.0.7
imageio==2.9.0
matplotlib==3.2.2
numpy==1.21.6
opencv_python==4.6.0.66
pandas==1.3.5
PyWavelets==1.4.1
scikit_image==0.18.3
scikit_learn==1.2.0
scipy==1.7.3
skimage==0.0
tqdm==4.64.1
detectron2.egg==info
torch==1.13.0+cu116

